refactor: extract extension.ts into command and service modules
Break the 2,021-line activate() god function into focused modules:

- commands/branchCommands.ts: 10 branch action handlers
- commands/commitCommands.ts: 13 commit context actions
- services/diffService.ts: file comparison and patch operations
- services/gitHelpers.ts: shared git utilities (validation, resolution)
- services/jetbrainsMergeService.ts: JetBrains merge tool orchestration
- services/refreshService.ts: debounced refresh + file watchers

extension.ts reduced from 2,021 to 518 lines (75% reduction).
Each extracted module is under 400 lines.
