+++ src/services/jetbrainsMergeService.ts
@@ -1,0 +1,253 @@
+// JetBrains merge tool orchestration extracted from extension.ts.
+// Handles configuration prompts, detection, launch, and post-merge
+// file inspection for JetBrains IDE merge tools.
+
+import * as fs from "fs";
+import * as path from "path";
+import * as vscode from "vscode";
+import { GitOps } from "../git/operations";
+import { getErrorMessage } from "../utils/errors";
+import {
+    containsConflictMarkers,
+    detectInstalledJetBrainsMergeToolCandidates,
+    detectInstalledJetBrainsMergeToolPath,
+    launchJetBrainsMergeTool,
+    resolveJetBrainsMergeBinaryPath,
+} from "../utils/jetbrainsMergeTool";
+import { runWithNotificationProgress } from "../utils/notifications";
+
+function getIntelliGitConfig(): vscode.WorkspaceConfiguration | null {
+    const getConfiguration = vscode.workspace.getConfiguration;
+    if (typeof getConfiguration !== "function") return null;
+    return getConfiguration.call(vscode.workspace, "intelligit");
+}
+
+export function getJetBrainsMergeToolPath(): string {
+    return getIntelliGitConfig()?.get<string>("jetbrainsMergeTool.path", "").trim() ?? "";
+}
+
+export function getPreferExternalMergeTool(): boolean {
+    return (
+        getIntelliGitConfig()?.get<boolean>("jetbrainsMergeTool.preferExternal", true) ?? true
+    );
+}
+
+function getDefaultJetBrainsMergeToolPath(): string {
+    switch (process.platform) {
+        case "darwin":
+            return "/Applications/PyCharm.app";
+        case "win32":
+            return "C:\\Program Files\\JetBrains\\PyCharm\\bin\\pycharm64.exe";
+        default:
+            return "pycharm";
+    }
+}
+
+async function saveJetBrainsMergeToolPath(rawPath: string): Promise<string | null> {
+    const trimmed = rawPath.trim();
+    if (!trimmed) return null;
+
+    if (path.isAbsolute(trimmed) && !fs.existsSync(trimmed)) {
+        vscode.window.showErrorMessage(`JetBrains path not found: ${trimmed}`);
+        return null;
+    }
+
+    let resolvedBinaryPath: string;
+    try {
+        resolvedBinaryPath = await resolveJetBrainsMergeBinaryPath(trimmed);
+    } catch (err) {
+        const msg = getErrorMessage(err);
+        vscode.window.showErrorMessage(`Invalid JetBrains merge tool path: ${msg}`);
+        return null;
+    }
+
+    const config = getIntelliGitConfig();
+    if (config && typeof config.update === "function") {
+        await config.update(
+            "jetbrainsMergeTool.path",
+            trimmed,
+            vscode.ConfigurationTarget.Global,
+        );
+    }
+
+    const resolutionText =
+        resolvedBinaryPath === trimmed
+            ? `Executable: ${resolvedBinaryPath}`
+            : `Resolved executable: ${resolvedBinaryPath}`;
+    vscode.window.showInformationMessage(`Saved JetBrains merge tool path. ${resolutionText}`);
+    return trimmed;
+}
+
+export async function promptForJetBrainsMergeToolPath(): Promise<string | null> {
+    const existing = getJetBrainsMergeToolPath();
+    const detected = existing ? null : await detectInstalledJetBrainsMergeToolPath();
+    const suggested = existing || detected || getDefaultJetBrainsMergeToolPath();
+    const input = await vscode.window.showInputBox({
+        title: "JetBrains Merge Tool Path",
+        prompt: "Enter a JetBrains IDE binary path/command (pycharm, idea, webstorm) or a macOS .app bundle path.",
+        placeHolder: suggested,
+        value: suggested,
+        ignoreFocusOut: true,
+    });
+    if (!input) return null;
+
+    return saveJetBrainsMergeToolPath(input);
+}
+
+export async function detectAndPickJetBrainsMergeToolPath(): Promise<string | null> {
+    const candidates = await detectInstalledJetBrainsMergeToolCandidates();
+    if (candidates.length === 0) {
+        vscode.window.showWarningMessage(
+            "No JetBrains IDE installations were auto-detected. Enter the path manually instead.",
+        );
+        return promptForJetBrainsMergeToolPath();
+    }
+
+    const quickPickItems = await Promise.all(
+        candidates.slice(0, 50).map(async (candidatePath) => {
+            let detail: string | undefined;
+            try {
+                const resolved = await resolveJetBrainsMergeBinaryPath(candidatePath);
+                detail = resolved === candidatePath ? undefined : `Resolved: ${resolved}`;
+            } catch {
+                detail = undefined;
+            }
+            return {
+                label: path.basename(candidatePath),
+                description: candidatePath,
+                detail,
+                candidatePath,
+            };
+        }),
+    );
+
+    const MANUAL_SENTINEL = "__manual__";
+    quickPickItems.push({
+        label: "$(edit) Enter path manually",
+        description: "Open the path prompt",
+        detail: undefined,
+        candidatePath: MANUAL_SENTINEL,
+    });
+
+    const picked = await vscode.window.showQuickPick(quickPickItems, {
+        title: "Detect JetBrains Merge Tool",
+        placeHolder: "Select a detected JetBrains IDE to use as the merge tool",
+        ignoreFocusOut: true,
+        matchOnDescription: true,
+        matchOnDetail: true,
+    });
+    if (!picked) return null;
+    if (picked.candidatePath === MANUAL_SENTINEL) {
+        return promptForJetBrainsMergeToolPath();
+    }
+    return saveJetBrainsMergeToolPath(picked.candidatePath);
+}
+
+function sleep(ms: number): Promise<void> {
+    return new Promise((resolve) => {
+        setTimeout(resolve, ms);
+    });
+}
+
+async function readMergedFileWithRetry(
+    outputFileFsPath: string,
+    beforeMergeText: string | null,
+): Promise<string> {
+    let lastReadError: unknown;
+    const delaysMs = [0, 80, 160, 320, 500];
+
+    for (let attempt = 0; attempt < delaysMs.length; attempt++) {
+        if (delaysMs[attempt] > 0) {
+            await sleep(delaysMs[attempt]);
+        }
+
+        try {
+            const text = await fs.promises.readFile(outputFileFsPath, "utf8");
+            const unchanged = beforeMergeText !== null && text === beforeMergeText;
+            const hasConflictBlock = containsConflictMarkers(text);
+            if ((unchanged || hasConflictBlock) && attempt < delaysMs.length - 1) {
+                continue;
+            }
+            return text;
+        } catch (readErr) {
+            lastReadError = readErr;
+            if (attempt === delaysMs.length - 1) throw readErr;
+        }
+    }
+
+    throw lastReadError instanceof Error
+        ? lastReadError
+        : new Error("Failed to read merged file after external merge tool closed.");
+}
+
+export async function openJetBrainsMergeToolForFile(
+    filePath: string,
+    repoRoot: string,
+    gitOps: GitOps,
+    refreshConflictUi: () => Promise<void>,
+    openBuiltInMergeEditorForFile: (filePath: string) => Promise<void>,
+): Promise<boolean> {
+    let jetBrainsPath = getJetBrainsMergeToolPath();
+    if (!jetBrainsPath) {
+        const action = await vscode.window.showInformationMessage(
+            "JetBrains merge tool path is not configured.",
+            "Configure",
+            "Open VS Code Merge Editor",
+        );
+        if (action === "Open VS Code Merge Editor") {
+            await openBuiltInMergeEditorForFile(filePath);
+            return true;
+        }
+        if (action !== "Configure") return false;
+        const configured = await promptForJetBrainsMergeToolPath();
+        if (!configured) return false;
+        jetBrainsPath = configured;
+    }
+
+    try {
+        const versions = await gitOps.getConflictFileVersions(filePath);
+        const outputFileFsPath = path.join(repoRoot, filePath);
+        const beforeMergeText = await fs.promises
+            .readFile(outputFileFsPath, "utf8")
+            .catch(() => null);
+
+        await runWithNotificationProgress(
+            `Opening JetBrains merge tool for ${filePath}...`,
+            async () => {
+                await launchJetBrainsMergeTool({
+                    binaryPath: jetBrainsPath,
+                    repoRootFsPath: repoRoot,
+                    relativeFilePath: filePath,
+                    outputFileFsPath,
+                    baseContent: versions.base,
+                    oursContent: versions.ours,
+                    theirsContent: versions.theirs,
+                });
+            },
+        );
+
+        try {
+            const mergedText = await readMergedFileWithRetry(outputFileFsPath, beforeMergeText);
+            if (!containsConflictMarkers(mergedText)) {
+                await gitOps.stageFile(filePath);
+                vscode.window.showInformationMessage(`Merged and staged: ${filePath}`);
+            } else {
+                vscode.window.showInformationMessage(
+                    `Merge tool closed, but conflict markers remain in ${filePath}`,
+                );
+            }
+        } catch (readErr) {
+            const msg = getErrorMessage(readErr);
+            vscode.window.showWarningMessage(
+                `Could not inspect merged file '${filePath}' after JetBrains merge: ${msg}`,
+            );
+        }
+
+        await refreshConflictUi();
+        return true;
+    } catch (error) {
+        const message = getErrorMessage(error);
+        vscode.window.showErrorMessage(`JetBrains merge tool failed: ${message}`);
+        return false;
+    }
+}
